refactor(cli): remove redundant branded scalar detection system
The branded-detector, scalar-registry, and symbol-resolver modules were
redundant because standard scalar types (IDString, Int, Float, IDNumber)
are defined using DefineScalar and can be detected by detectScalarMetadata.

Changes:
- Remove branded-detector.ts, scalar-registry.ts, symbol-resolver.ts
- Unify scalar detection using detectScalarMetadata only
- Rename brandName to typeName in ScalarTypeInfo interface
- Update comments to use "metadata" instead of "branded type"
