+++ packages/cli/src/resolver-extractor/validator/only-validator.test.ts
@@ -21,7 +21,7 @@ function createScalarTypeRef(
     nullable: false,
     scalarInfo: {
       scalarName,
-      brandName: scalarName,
+      typeName: scalarName,
       baseType: undefined,
       isCustom: true,
       only,
