+++ packages/cli/src/type-extractor/types/index.ts
@@ -1,4 +1,3 @@
-export type { ScalarTypeInfo } from "../../shared/branded-detector.js";
 export type { DeprecationInfo } from "../../shared/tsdoc-parser.js";
 export type {
   Diagnostic,
@@ -17,6 +16,7 @@ export type {
   EnumMemberInfo,
   ExtractedTypeInfo,
   FieldDefinition,
+  ScalarTypeInfo,
   TSTypeReference,
   TypeKind,
   TypeMetadata,
